[Backout] - CL31811875 [FYI] michael.bao Original CL Desc ----------------------------------------------------------------- increase maximum value contained by FImplicitObjectUnion::NumLeafObjects from uint16 to int32
#rb Chris.Caulfield

[CL 31817439 by michael bao in ue5-main branch]

+++ Engine/Source/Runtime/Experimental/Chaos/Private/Chaos/ImplicitObjectUnion.cpp
@@ -160,7 +160,7 @@ void FImplicitObjectUnion::SetNumLeafObjects(int32 InNumLeafObjects)
 	ensure(InNumLeafObjects <= MaxNumLeafObjects);
 	check(InNumLeafObjects >= 0);
 
-	NumLeafObjects = FMath::Min(InNumLeafObjects, MaxNumLeafObjects);
+	NumLeafObjects = uint16(FMath::Min(InNumLeafObjects, MaxNumLeafObjects));
 }
 
 void FImplicitObjectUnion::CreateBVH()
